requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosmtplib==3.0.2
aiosqlite==0.21.0
alembic==1.15.2
alembic-postgresql-enum==1.7.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.9
async-timeout==5.0.1
babel==2.17.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
dill==0.3.9
Django==5.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-admin==1.0.4
fastapi-cli==0.0.7
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
orjson==3.10.16
passlib==1.7.4
pendulum==3.0.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.1
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.6
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
slugify==0.0.1
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.46.1
time-machine==2.16.0
tomlkit==0.13.2
tortoise-orm==0.25.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
